Minor style changes
Changed the nav-links buttons in a form inside adminNavbar.jsp to anchors in a generic div with the same class.
Changed the includo of the adminNavbar.jsp in adminHeader.jsp.
Added a paragraph to adminAccount.jsp to test the charset.

Renamed the javascript folder in scripts.

Added a class logo-large for a big logo in formHeader.jsp.
Edited index.jsp and catalogo.jsp to have a bigger add-to-cart icon.
Renamed the links class in navbar.jsp to link.

Edited catalog.css with minor changes to the product box style.
Edited header.css with minor changes to the logo, searchbar and navbar style.
Edited page.css to add the use of the css color variables.
Edited user.css with minor changes to the form page style.
Edited light.css to change the value of a color variable.