Added a page favicon and minor style changes
Edited header.jsp to add the link to the icon picture for the favicon.
Added the website logo icon.

Edited light.css to add test color variables for the second palette.
Edited page.css to change the hr elements' margin.
Edited user.css to style the error paragraphs in the forms and the form box margins.