Added order and user js with page and style integration
Edited page.css with new style for a save button and not displayed elements.

Added order.js with functions to use in the order page.
Edited order.css to change the style of some elements.
Edited orderPage.jsp to adapt the page to the changes.

Edited user.js with functions to use in the user forms.
Edited user.css to change the style of some elements.
Edited userData.jsp, userAddresses.jsp and userPayMethods.jsp to adapt the pages to the changes.

Updated admin.js with the new changes.